pyproject.toml
[project]
name = "camp"
version = "2024.0.0"
requires-python = ">= 3.11"
dependencies = [
  "altair[all]==5.5.0",
  "fsspec==2025.7.0",
  "keras==3.11.2",
  "manim==0.19.0",
  "matplotlib==3.10.5",
  "minio==7.2.16",
  "numpy==2.3.2",
  "pandas==2.3.1",
  "polars==1.32.2",
  "psutil==7.0.0",
  "pyarrow==21.0.0",
  "rich[jupyter]==14.1.0",
  "s3fs==2025.7.0",
  "scikit-learn==1.7.1",
  "scipy==1.16.1",
  "seaborn==0.13.2",
  "statsmodels==0.14.5",
  "streamlit==1.48.1",
  "sympy==1.14.0",
]

[project.optional-dependencies]
distributed = [
  "pyspark==4.0.1",
]
gym = [
  "gymnasium[atari,box2d,classic-control,mujoco,toy-text]==1.2.0",
]
tensorflow = [
  "tensorflow==2.20.0",
]
text = [
  "gensim==4.3.3",
  "nltk==3.9.1",
  "spacy==3.8.7",
]
torch = [
  "torch==2.8.0",
  "torchvision==0.23.0",
]
torch-common = [
  "deepspeed @ https://github.com/yjoer/camp/releases/download/v2025.6.0/deepspeed-0.15.0+55b4cae8-py3-none-any.whl",
  "safetensors==0.6.2",
  "torchmetrics==1.8.1",
]
torch-cu129 = [
  "torch==2.8.0",
  "torchvision==0.23.0",
]
tree = [
  "catboost==1.2.8",
  "lightgbm==4.6.0",
  "xgboost==3.0.4",
]
vision = [
  "mmcv>=2.2.0",
  "mmdet>=3.3.0",
  "mmengine>=0.10.4",
  "motmetrics @ git+https://github.com/cheind/py-motmetrics@057504b",
  "pillow==11.3.0",
  "supervision>=0.22.0",
  "timm==1.0.19",
  "ultralytics==8.3.189",
]

[build-system]
requires = ["maturin>=1,<2"]
build-backend = "maturin"

[tool.jupytext]
preferred_jupytext_formats_read = "cpp:percent,py:percent,rs:percent"
notebook_metadata_filter = "-all"

[tool.marimo.display]
custom_css = ["marimo.css"]
code_editor_font_size = 16

[tool.marimo.runtime]
pythonpath = ["."]

[tool.maturin]
features = ["pyo3/extension-module"]
module-name = "camp.rs"

[tool.mypy]
cache_dir = ".venv/.mypy_cache"
explicit_package_bases = true
ignore_missing_imports = true
allow_redefinition = true

[tool.pytest.ini_options]
addopts = "-n auto"
cache_dir = ".venv/.pytest_cache"
python_files = "*.py"
testpaths = ["camp", "solutions"]

[tool.ruff]
cache-dir = ".venv/.ruff_cache"
extend = "node_modules/@camp/config/ruff.toml"

[tool.ruff.lint.per-file-ignores]
"notebooks/**.py" = [
  "B015",  # useless-comparison
  "B018",  # useless-expression
]
"solutions/**.py" = [
  "B018",  # useless-expression
]

[tool.uv]
managed = true
build-constraint-dependencies = ["setuptools==77.0.3"]
conflicts = [
  [
    { extra = "torch" },
    { extra = "torch-cu129" },
  ],
]
dev-dependencies = [
  "black==25.1.0",
  "isort==6.0.1",
  "jedi-language-server==0.45.1",
  "jupyter-collaboration==4.1.1",
  "jupyterlab-code-formatter==3.0.2",
  "jupyterlab-git==0.51.2",
  "jupyterlab-lsp==5.2.0",
  "jupyterlab-quickopen @ https://github.com/yjoer/camp/releases/download/v2025.6.0/jupyterlab_quickopen-2.0.1-py3-none-any.whl",
  "jupyterlab==4.5.0b0",
  "jupytext==1.17.3",
  "marimo==0.16.2",
  "maturin==1.9.3",
  "mypy==1.17.1",
  "notebook-intelligence==2.4.2",
  "pytest-xdist==3.8.0",
  "pytest==8.4.1",
  "ruff==0.13.2",
  "scalene==1.5.53",
]

[tool.uv.workspace]
members = ["examples/grafana/*"]

[[tool.uv.dependency-metadata]]
name = "gensim"
version = "4.3.3"
requires-dist = ["scipy>=1.14.0", "smart-open>=1.8.1"]

[tool.uv.sources]
torch = [
  { index = "pytorch", extra = "torch" },
  { index = "pytorch-cu129", extra = "torch-cu129" },
]
torchvision = [
  { index = "pytorch", extra = "torch" },
  { index = "pytorch-cu129", extra = "torch-cu129" },
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu129"
url = "https://download.pytorch.org/whl/cu129"
explicit = true
